% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andorR-package.R
\docType{package}
\name{andorR-package}
\alias{andorR}
\alias{andorR-package}
\title{andorR: An Analysis and Optimisation Tool for AND-OR Decision Trees}
\description{
The \code{andorR} package provides a suite of tools to create, analyze, and
interactively solve complex logical decision trees. It is designed for
problems where a final TRUE/FALSE conclusion is determined by propagating
answers and their associated confidence scores up a hierarchical structure
of AND/OR rules. The package's core feature is an optimization algorithm that
guides the user to the most influential questions, minimizing the effort
required to reach a confident conclusion.
}
\section{Key Functions}{

The main workflow is built around a few key functions:
\itemize{
\item \code{\link{load_tree_csv}}, \code{\link{load_tree_df}},
\code{\link{load_tree_yaml}}, \code{\link{load_tree_node_list}},
\code{\link{load_tree_df_path}}, \code{\link{load_tree_csv_path}}: Load
your decision tree from different formats.
\item \code{\link{set_answer}}: Answer a question and provide a confidence
score.
\item \code{\link{update_tree}}: The core calculation engine that initializes
and recalculates all logical states and influence indices.
\item \code{\link{get_highest_influence}},
\code{\link{get_confidence_boosters}}: Prioritisation of questions to
optimise the completion of the tree.
\item \code{\link{print_tree}} and \code{\link{print_questions}}: Visualise
the state of the tree
\item \code{\link{andorR_interactive}}: The main, user-facing function that
automates the entire analysis in a step-by-step interactive session.
}
}

\section{Full Tutorials (Vignettes)}{

To learn how to use the package in detail, please see the vignettes:
#' To learn how to use the package in detail, please see the vignettes:
\describe{
\item{Introduction:}{
\code{vignette("introduction-to-andorR", package = "andorR")}
}
\item{Explanation of data formats:}{
\code{vignette("data-formats", package = "andorR")}
}
\item{Tree optimisation algorithm:}{
\code{vignette("tree-optimisation", package = "andorR")}
}
\item{Sensitivity analysis algorithm:}{
\code{vignette("confidence-boosting", package = "andorR")}
}
\item{Example files:}{
\code{vignette("example-data-files", package = "andorR")}
}
}
#' @keywords internal
}

\author{
\strong{Maintainer}: Angus R Cameron \email{angus.cameron@epimundi.com} (\href{https://orcid.org/0000-0001-8801-0366}{ORCID})

}
