% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{load.tree.csv}
\alias{load.tree.csv}
\title{Load a decision tree from a CSV file}
\usage{
load.tree.csv(file_path)
}
\arguments{
\item{file_path}{The path to the .csv file. The file is expected to have
columns: id, name, question, rule, parent.}
}
\value{
A `data.tree` object.
}
\description{
A wrapper function that reads a CSV file into a data frame
  and then uses `load.tree.df` to construct the tree object.
}
\examples{
\dontrun{
# Assuming "my_decision_tree.csv" is in the working directory
my_tree <- load.tree.csv("my_decision_tree.csv")
}
}
