% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{load.tree.df}
\alias{load.tree.df}
\title{Build the decision tree from a data frame}
\usage{
load.tree.df(df)
}
\arguments{
\item{df}{A data frame with columns: id, name, question, rule, parent.}
}
\value{
A data.tree object.
}
\description{
Takes a data frame with a specific parent-child structure and
converts it into a data.tree object, initializing necessary attributes.
}
