% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{load.tree.yaml}
\alias{load.tree.yaml}
\title{Load a decision tree from a YAML file}
\usage{
load.tree.yaml(file_path)
}
\arguments{
\item{file_path}{The path to the .yml or .yaml file.}
}
\value{
A `data.tree` object.
}
\description{
A wrapper function that reads a YAML file into a nested list
  and then uses `load.tree.node.list` to construct the tree object.
}
\examples{
\dontrun{
# Assuming "my_decision_tree.yml" is in the working directory
my_tree <- load.tree.yaml("my_decision_tree.yml")
}
}
