% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{load_tree_csv}
\alias{load_tree_csv}
\title{Load a decision tree from a CSV file}
\usage{
load_tree_csv(file_path)
}
\arguments{
\item{file_path}{The path to the .csv file. The file is expected to have
columns: id, name, question, rule, parent.}
}
\value{
A \code{data.tree} object.
}
\description{
A wrapper function that reads a CSV file into a data frame
and then uses \code{load_tree_df} to construct the tree object.
}
\examples{
\dontrun{
# Assuming "my_decision_tree.csv" is in the working directory
my_tree <- load_tree_csv("my_decision_tree.csv")
}
}
