% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{load_tree_df}
\alias{load_tree_df}
\title{Build the decision tree from a data frame}
\usage{
load_tree_df(df)
}
\arguments{
\item{df}{A data frame with columns: id, name, question, rule, parent.}
}
\value{
A data.tree object.
}
\description{
Takes a data frame with a specific parent-child structure and
converts it into a data.tree object, initializing necessary attributes.
}
\examples{
# Load a tree from the 'ethical' dataframe included in this package
ethical_tree <- load_tree_df(ethical)

# View the tree structure
print_tree(ethical_tree)

}
