% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{load_tree_yaml}
\alias{load_tree_yaml}
\title{Load a decision tree from a YAML file}
\usage{
load_tree_yaml(file_path)
}
\arguments{
\item{file_path}{The path to the .yml or .yaml file.}
}
\value{
A \code{data.tree} object.
}
\description{
A wrapper function that reads a YAML file into a nested list
and then uses \code{load_tree_node_list} to construct the tree object.
}
\examples{
\dontrun{
#' # Load data from the `ethical.yml` file included with this package
path <- system.file("extdata", "ethical.yml", package = "andorR")
ethical_tree <- load.tree.yaml(path)

# View the tree
print_tree(ethical_tree)
}
}
