% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{notifiable}
\alias{notifiable}
\title{Notifiable Aquatic Animal Disease decision tree}
\format{
A data frame with...
}
\usage{
notifiable
}
\description{
Notifiable Aquatic Animal Disease decision tree
}
\keyword{datasets}
