% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tree.R
\name{validate_tree_df}
\alias{validate_tree_df}
\title{Validate the structure of a relational tree data frame}
\usage{
validate_tree_df(df)
}
\arguments{
\item{df}{The data frame to validate.}
}
\value{
Returns \code{TRUE} if the data frame is valid, otherwise it stops with
a descriptive error message.
}
\description{
Checks if a data frame has the correct columns, data types,
and structural integrity to be converted into a valid decision tree.
}
\keyword{internal}
