% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tree.R
\name{validate_tree_df_path}
\alias{validate_tree_df_path}
\title{Validate the structure of a path-string tree data frame}
\usage{
validate_tree_df_path(df, delim = "/")
}
\arguments{
\item{df}{The data frame to validate.}

\item{delim}{The character used to separate nodes in the path string.}
}
\value{
Returns \code{TRUE} if the data frame is valid, otherwise it stops with
a descriptive error message.
}
\description{
Checks if a data frame in path-string format has the correct
columns, data types, and structural integrity to be converted into a
valid decision tree.
}
\keyword{internal}
