% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ethical}
\alias{ethical}
\title{Ethical Investment Tree for a Fictional Company}
\format{
A data frame with...
}
\usage{
ethical
}
\description{
Ethical Investment Tree for a Fictional Company
}
\keyword{datasets}
