% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{notifiable}
\alias{notifiable}
\title{Notifiable aquatic animal disease decision tree}
\format{
A data frame with 5 variables and 50 rows

\describe{
Each rows represents a node or leaf in the tree and the columns represent
attributes of those nodes. The columns are:
\item{id}{A unique sequential numeric identifier for each node}
\item{name}{A short, unique alphanumeric code or name for nodes. For
leaf nodes (questions), a short code is used. For higher nodes, a
descriptive phrase is used.}
\item{question}{The full text of the question for leaves, or NA for
higher nodes.}
\item{rule}{The logical rule for nodes, either \strong{AND} or \strong{OR}, and
NA for leaves.}
\item{parent}{The numeric id of the parent node, and NA for the root node.}

}
}
\source{
This example is based on a decision tree developed to guide provincial
governments on the Canada Atlantic coast in their decisions on the
listing or de-listing of notifiable aquatic animal diseases.
}
\usage{
notifiable
}
\description{
This dataframe represents a decision tree in relational format, in
which hierarchical relationships are indicated by a value indicating
the parent of each node.

The decision tree is based on a tool to standardise the process of
determining if an aquatic animal disease should be made officially
notifiable to the veterinary authorities.
}
\details{
A \code{data.tree} object is created from the dataframe using the \code{read_tree_df()}
function.
}
\examples{
# Load the tree definition from the dataframe into a data.tree object
#   ready for subsequent analysis
tree <- load_tree_df(notifiable)

# View the tree
print_tree(tree)

}
\keyword{datasets}
