% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{update_tree}
\alias{update_tree}
\title{Update a Tree Based on Answers Provided}
\usage{
update_tree(tree)
}
\arguments{
\item{tree}{The `data.tree` object to be modified.}
}
\value{
Returns the modified `tree` object invisibly, which allows for function chaining.
}
\description{
Propagate the results up to the tree nodes based on the answers provided,
and update the influence index to identify most important questions.
}
